🦀 Use ferris as the default ship
Offset the tiled objects into Bevy game coordinates

Refactor game state from main into GameLogicPlugin
