+++ spacerace-client/src/main.rs
@@ -1,11 +1,12 @@
 use crossterm::event::{self, Event, KeyCode};
-use crossterm::{cursor, execute, terminal};
+use crossterm::{cursor, execute, terminal, queue, };
 use rand::distributions::DistString;
 use rand::Rng;
 use serde::{Deserialize, Serialize};
 use std::env;
 use std::io::{stdout, Write};
 use std::time::Duration;
+use reqwest::Client;
 use tokio::time::sleep;
 use tracing::{debug, error, info};
 use tracing_subscriber;
@@ -44,6 +45,14 @@ enum GameStateResponse {
     Active(ActiveGameState),
 }
 
+#[derive(Debug, Deserialize, PartialEq)]
+enum GameStatus {
+    Queued,
+    Running,
+    Finished,
+}
+
+
 /// Details of the currently active game state.
 #[derive(Debug, Deserialize)]
 struct ActiveGameState {
@@ -53,8 +62,8 @@ struct ActiveGameState {
     ships: Vec<ShipState>,
     /// Name of the current map.
     map_name: String,
-    /// Textual representation of the game state ("Active", "Queued", etc.)
-    state: String,
+    /// Game state ("Active", "Queued", etc.)
+    state: GameStatus,
 }
 
 /// Represents the state of a single ship.
@@ -92,7 +101,7 @@ struct ControlResponse {
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    // Set up logging with INFO level by default.
+    // Set up logging
     tracing_subscriber::fmt()
         .with_max_level(tracing::Level::INFO)
         .init();
@@ -103,11 +112,72 @@ async fn main() -> anyhow::Result<()> {
     let player_password = env::var("PLAYER_PASSWORD").unwrap_or_else(|_| {
         rand::distributions::Alphanumeric.sample_string(&mut rand::thread_rng(), 16)
     });
-
-    let player_team = player_name.clone(); // Use player's name as team name.
+    let player_team = env::var("PLAYER_TEAM").unwrap_or_else(|_|player_name.clone()); // Default to use player's name as team name.
 
     let client = reqwest::Client::new();
+    let lobby_response = register_player(&client, &host, player_name, player_team, &player_password).await?;
 
+    info!("Registered with lobby: {:?}", lobby_response);
+
+    // Set up terminal for keyboard input
+    crossterm::terminal::enable_raw_mode()?;
+    let mut stdout = stdout();
+    execute!(stdout, cursor::Hide)?;
+
+
+
+    loop {
+        // Initial instructions
+        execute!(
+            stdout,
+            terminal::Clear(terminal::ClearType::All),
+            cursor::MoveTo(0, 0)
+        )?;
+
+        stdout.flush()?;
+
+        let state_response = get_server_state(&client, &host).await?;
+
+        match state_response {
+            GameStateResponse::Inactive => {
+                println!("Waiting for next game to start.");
+            }
+            GameStateResponse::Active(active) => {
+                // Check if the active game is the one we have joined
+                if active.game_id == lobby_response.game_id {
+                    // Great, *our* game is active.
+                    println!("Our game ({}) is now active. Starting game loop...", active.game_id);
+
+                    run_game_loop(
+                        client.clone(),
+                        player_password.clone(),
+                        host.clone(),
+                    ).await?;
+
+                    // If `run_game_loop` returns, that means the game ended or the user quit.
+                    // Break out of the outer loop to exit, or re-register for another game, etc.
+                    println!("Game loop ended. Exiting outer loop...");
+                    break;
+                } else {
+                    // Another game is active, but not ours. We keep waiting, hopefully we haven't missed our game!
+                    println!(
+                        "Game '{}' is active, but we're in game '{}'. Waiting...",
+                        active.game_id,
+                        lobby_response.game_id
+                    );
+                    sleep(Duration::from_secs(1)).await;
+                }
+
+            }
+        }
+    }
+    // Run the game loop
+    run_game_loop(client, player_password, host).await?;
+
+    Ok(())
+}
+
+async fn register_player(client: &Client, host: &String, player_name: String, player_team: String, player_password: &String) -> anyhow::Result<LobbyResponse> {
     // Register with the lobby
     let lobby_req = LobbyRequest {
         name: player_name.clone(),
@@ -125,25 +195,7 @@ async fn main() -> anyhow::Result<()> {
         .await?;
 
     info!("Registered with lobby: {:?}", lobby_response);
-
-    // Set up terminal for keyboard input
-    crossterm::terminal::enable_raw_mode()?;
-    let mut stdout = stdout();
-    execute!(stdout, cursor::Hide)?;
-
-    // Initial instructions
-    execute!(
-        stdout,
-        terminal::Clear(terminal::ClearType::All),
-        cursor::MoveTo(0, 0)
-    )?;
-    info!("Use arrow keys (or WASD) to control the ship. Press 'q' to quit.");
-    stdout.flush()?;
-
-    // Run the game loop
-    run_game_loop(client, player_password, host).await?;
-
-    Ok(())
+    Ok(lobby_response)
 }
 
 /// Run the main game loop:
@@ -165,14 +217,7 @@ async fn run_game_loop(
     let mut stdout = stdout();
 
     loop {
-        // Get current game state
-        let state_response = client
-            .get(format!("{}/state", server_url))
-            .send()
-            .await?
-            .error_for_status()?
-            .json::<GameStateResponse>()
-            .await?;
+        let state_response = get_server_state(&client, &server_url).await?;
 
         // Clear the screen each loop and move cursor to top-left
         execute!(
@@ -181,24 +226,19 @@ async fn run_game_loop(
             cursor::MoveTo(0, 0)
         )?;
 
-        // Reprint instructions so they stay visible
-        println!("Use arrow keys (or WASD) to control the ship. Press 'q' to quit.");
-
         // Determine thrust and rotation based on user input
         let thrust_rotation = if let GameStateResponse::Active(active) = &state_response {
-            if active.state != "Active" && active.state != "Queued" {
-                debug!(
-                    "Game is not in an active state (state: {}). Press 'q' to quit.",
-                    active.state
-                );
-                thrust_rotation_from_input_with_timeout(Duration::from_millis(200))?
+            if active.state == GameStatus::Finished {
+                debug!("Game finished!");
+                return Ok(());
+
             } else {
                 // Game is active or queued, allow user to control
                 thrust_rotation_from_input_with_timeout(Duration::from_millis(200))?
             }
         } else {
-            debug!("No active game. Press 'q' to quit.");
-            thrust_rotation_from_input_with_timeout(Duration::from_millis(200))?
+            debug!("No active game. Waiting for next game to start.");
+            None
         };
 
         // If user provided control input, send it to server
@@ -225,7 +265,7 @@ async fn run_game_loop(
         if let GameStateResponse::Active(active) = &state_response {
             for ship in &active.ships {
                 println!(
-                    "Ship ID: {}, Position: ({:.1}, {:.1}), Velocity: ({:.1}, {:.1}), Orientation: {:.1}",
+                    "Ship ID: {}, Position: ({:.3}, {:.3}), Velocity: ({:.3}, {:.3}), Orientation: {:.3}",
                     ship.id,
                     ship.position[0],
                     ship.position[1],
@@ -236,10 +276,27 @@ async fn run_game_loop(
             }
         }
 
+        // Reprint instructions so they stay visible
+        println!("Use arrow keys (or WASD) to control the ship. Press 'q' to quit.");
+
         stdout.flush()?;
+
         // Throttle updates slightly to avoid high CPU usage
         sleep(Duration::from_millis(30)).await;
     }
+}
+
+async fn get_server_state(client: &Client, host: &String) -> anyhow::Result<GameStateResponse> {
+    // Get current game state
+    let state_response = client
+        .get(format!("{}/state", host))
+        .send()
+        .await?
+        .error_for_status()?
+        .json::<GameStateResponse>()
+        .await?;
+
+    Ok(state_response)
 }
 
 /// Polls for keyboard input within the given timeout and returns thrust/rotation commands.
