+++ spacerace-server/src/game_logic/mod.rs
@@ -2,24 +2,14 @@ use crate::app_state::AppState;
 use crate::components::ship::ControllableShip;
 use crate::components::ship::Ship;
 use crate::game_state::{GameState, GameStatus};
+use crate::{components, game_state};
 use bevy::prelude::*;
 use bevy_rapier2d::prelude::*;
+use rand::prelude::SliceRandom;
+use rand::Rng;
 use std::time::Duration;
 
 pub struct GameLogicPlugin;
-
-impl Plugin for GameLogicPlugin {
-    fn build(&self, app: &mut App) {
-        app.add_systems(
-            PostUpdate,
-            transition_to_inactive_system.run_if(has_timer_and_game_finished),
-        )
-        .add_systems(
-            OnExit(ServerState::Active),
-            (cleanup_finished_game, cleanup_transition_timer),
-        );
-    }
-}
 
 // Enum that will be used as a global state for the game server
 #[derive(Clone, Copy, Default, Eq, PartialEq, Debug, Hash, States)]
@@ -27,6 +17,146 @@ pub enum ServerState {
     #[default]
     Inactive,
     Active,
+}
+
+impl Plugin for GameLogicPlugin {
+    fn build(&self, app: &mut App) {
+        app.init_state::<ServerState>()
+            .add_systems(
+                Update,
+                check_all_players_finished_system.run_if(in_state(ServerState::Active)),
+            )
+            .add_systems(
+                OnEnter(ServerState::Active),
+                (setup_scene, spawn_ships, start_game),
+            )
+            .add_systems(OnExit(ServerState::Active), (unload_game_entities))
+            .add_systems(OnEnter(ServerState::Inactive), (setup_game_scheduler))
+            .add_systems(
+                Update,
+                game_scheduler_system.run_if(in_state(ServerState::Inactive)),
+            )
+            .add_systems(
+                PostUpdate,
+                transition_to_inactive_system.run_if(has_timer_and_game_finished),
+            )
+            .add_systems(
+                OnExit(ServerState::Active),
+                (cleanup_finished_game, cleanup_transition_timer),
+            );
+    }
+}
+
+pub fn start_game(mut commands: Commands, app_state: Res<AppState>) {
+    let mut active_game_guard = app_state.active_game.lock().unwrap();
+    if let Some(active_game) = active_game_guard.as_mut() {
+        active_game.state = game_state::GameStatus::Running;
+    } else {
+        info!("No active game to start");
+    }
+}
+
+pub fn unload_game_entities(
+    mut commands: Commands,
+    query: Query<Entity, With<components::ActiveGameEntity>>,
+) {
+    // Despawn all entities from the game
+    for entity in query.iter() {
+        commands.entity(entity).despawn_recursive();
+    }
+}
+
+pub fn spawn_ships(
+    mut commands: Commands,
+    asset_server: Res<AssetServer>,
+    app_state: Res<AppState>
+) {
+    // Spawn a Ship for each player in the active GameState
+    let sprite_size = 25.0;
+    let mut rng = rand::thread_rng();
+
+    let active_game_guard = app_state.active_game.lock().unwrap();
+    if let Some(active_game) = active_game_guard.as_ref() {
+        for player in &active_game.players {
+            // Generate a random hue for this player's ship
+            let hue = rng.gen_range(0.0..360.0);
+            let color = Color::hsl(hue, 0.8, 0.5);
+
+            // Pick a random position for the ship from the map's start zones
+            let start_region = active_game.map.start_regions.choose(&mut rng).unwrap();
+
+            commands.spawn((
+                components::ActiveGameEntity,
+                components::ship::ControllableShip {
+                    id: player.id,
+                    impulse: 8_000.0,
+                    torque_impulse: 8_000.0,
+                },
+                Sprite {
+                    //color,
+                    image: asset_server.load("ferris.png"),
+                    custom_size: Some(Vec2::new(sprite_size, sprite_size)),
+                    ..Default::default()
+                },
+                // TODO sample position within the region's polygon
+                Transform::from_xyz(start_region.position.x, start_region.position.y, 0.0),
+                RigidBody::Dynamic,
+                Damping {
+                    linear_damping: 0.2,
+                    angular_damping: 0.5,
+                },
+                ExternalImpulse::default(),
+                AdditionalMassProperties::Mass(200.0),
+                Restitution::coefficient(0.9),
+                Friction::coefficient(0.5),
+                Collider::ball(sprite_size / 2.0),
+                Velocity::default(),
+                ActiveEvents::COLLISION_EVENTS,
+                ContactForceEventThreshold(10.0),
+            ));
+        }
+    } else {
+        info!("No active game to spawn ships for");
+    }
+}
+
+pub fn setup_scene(
+    mut commands: Commands,
+    mut rapier_config: Query<&mut RapierConfiguration>,
+    app_state: Res<AppState>,
+) {
+    if let Some(active_game) = app_state.active_game.lock().unwrap().as_ref() {
+        info!(game_id=?active_game.game_id, "Setting up scene for game");
+
+        let map = &active_game.map;
+
+        // Set up gravity using the map specific value
+        rapier_config.single_mut().gravity = Vec2::Y * map.gravity;
+
+        // Obstacles
+        for obstacle in &map.obstacles {
+            commands.spawn((
+                components::ActiveGameEntity,
+                Transform::from_xyz(obstacle.position.x, obstacle.position.y, 0.0),
+                Collider::polyline(obstacle.polygon.clone(), None),
+            ));
+        }
+
+        // Finish zone colliders
+        for finish in &map.finish_regions {
+            commands.spawn((
+                components::ActiveGameEntity,
+                Transform::from_xyz(finish.position.x, finish.position.y, 0.0),
+                // Note we only handle Tiled polygons, ideally handle rectangle objects etc
+                Collider::polyline(finish.polygon.clone(), None),
+                Sensor,
+                crate::components::FinishRegion,
+            ));
+        }
+    } else {
+        info!("No active game to set up scene for");
+        return;
+    }
 }
 
 // System to check if all players finished the race
